refactor: split types.TaskLogEntry into domain & ui
Move task log records and reporting projections into the domain package,
while keeping bubble tea list state and formatting in a UI-specific
adapter.

Persistence now works with domain models removing presentation
concerns from it.
